Xeon Phi Driver: Renaming and handling of IDs
Renamed xeon_phi.h to xeon_phi_internal.h
SMTP: reserved the top 8 entries for Xeon Phi mappings
      added functions to set the mappings based on the ID

Service: the service exchanges now the base addresses of the cards
         bugfix when waiting for the reply.

Interfaces: Renamed xeon_phi.if to xeon_phi_driver.if

+++ usr/drivers/xeon_phi/dma/dma_channel.h
@@ -69,8 +69,9 @@ struct xdma_channel
 
 static inline xeon_phi_dma_id_t xdma_chan_generate_id(struct xdma_channel *chan)
 {
+
     return ((((uint64_t) chan->chanid) << 56) | (((uint64_t) chan->head) << 32)
-            | (chan->reqcoutner++));
+            | (uint64_t)(chan->reqcoutner++));
 }
 
 /**
